add new block menu to group
